+++ packages/dropdown_button2/lib/src/utils.dart
@@ -25,6 +25,35 @@ double _clampDouble(double x, double min, double max) {
   return x;
 }
 
+/// Adds separators to a list of heights.
+///
+/// The [itemHeights] property is the list of heights of the items.
+///
+/// The [separatorHeight] property is the height of the separator.
+///
+/// Returns a new list of heights with separators added.
+List<double> addSeparatorsHeights({
+  required List<double> itemHeights,
+  required double? separatorHeight,
+}) {
+  final List<double> heights = [];
+
+  bool addSeparator = false;
+  if (separatorHeight != null) {
+    for (final item in itemHeights) {
+      if (addSeparator) {
+        heights.add(separatorHeight);
+      }
+      heights.add(item);
+      addSeparator = true;
+    }
+  } else {
+    heights.addAll(itemHeights);
+  }
+
+  return heights;
+}
+
 // ignore: public_member_api_docs
 extension ExtendedIterable<E> on Iterable<E> {
   /// Like Iterable<T>.map but the callback has index as second argument