+++ packages/dropdown_button2_test/lib/src/with_separators_example.dart
@@ -1,14 +1,14 @@
 import 'package:dropdown_button2/dropdown_button2.dart';
 import 'package:flutter/material.dart';
 
-class WithDividersExample extends StatefulWidget {
-  const WithDividersExample({super.key});
+class WithSeparatorsExample extends StatefulWidget {
+  const WithSeparatorsExample({super.key});
 
   @override
-  State<WithDividersExample> createState() => _WithDividersExampleState();
+  State<WithSeparatorsExample> createState() => _WithSeparatorsExampleState();
 }
 
-class _WithDividersExampleState extends State<WithDividersExample> {
+class _WithSeparatorsExampleState extends State<WithSeparatorsExample> {
   final List<String> items = [
     'Item1',
     'Item2',
@@ -16,37 +16,6 @@ class _WithDividersExampleState extends State<WithDividersExample> {
     'Item4',
   ];
   String? selectedValue;
-
-  List<DropdownItem<String>> _addDividersAfterItems(List<String> items) {
-    final List<DropdownItem<String>> menuItems = [];
-    for (final String item in items) {
-      menuItems.addAll(
-        [
-          DropdownItem<String>(
-            value: item,
-            height: 40,
-            child: Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 8.0),
-              child: Text(
-                item,
-                style: const TextStyle(
-                  fontSize: 14,
-                ),
-              ),
-            ),
-          ),
-          //If it's last item, we will not add Divider after it.
-          if (item != items.last)
-            const DropdownItem<String>(
-              enabled: false,
-              height: 4,
-              child: Divider(),
-            ),
-        ],
-      );
-    }
-    return menuItems;
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -62,7 +31,28 @@ class _WithDividersExampleState extends State<WithDividersExample> {
                 color: Theme.of(context).hintColor,
               ),
             ),
-            items: _addDividersAfterItems(items),
+            items: items
+                .map((String item) => DropdownItem<String>(
+                      value: item,
+                      height: 40,
+                      child: Padding(
+                        padding: const EdgeInsets.symmetric(horizontal: 8.0),
+                        child: Text(
+                          item,
+                          style: const TextStyle(
+                            fontSize: 14,
+                          ),
+                        ),
+                      ),
+                    ))
+                .toList(),
+            dropdownSeparator: const DropdownSeparator(
+              height: 4,
+              child: Padding(
+                padding: EdgeInsets.symmetric(horizontal: 8.0),
+                child: Divider(),
+              ),
+            ),
             value: selectedValue,
             onChanged: (String? value) {
               setState(() {