+++ packages/dropdown_button2_test/test/examples/with_separators_example_test.dart
@@ -7,15 +7,15 @@ import 'test_app.dart';
 
 void main() {
   runGoldenTests(
-    'With Dividers Example',
+    'With Separators Example',
     (WidgetTester tester, ViewVariant variant) async {
       // GIVEN
-      await tester.pumpTestApp(const WithDividersExample());
+      await tester.pumpTestApp(const WithSeparatorsExample());
 
       // THEN
       await expectLater(
         find.byType(TestApp),
-        matchesGoldenFile('goldens/with_dividers_example/${variant.name}_closed_menu.png'),
+        matchesGoldenFile('goldens/with_separators_example/${variant.name}_closed_menu.png'),
       );
 
       // GIVEN
@@ -26,7 +26,7 @@ void main() {
       // THEN
       await expectLater(
         find.byType(TestApp),
-        matchesGoldenFile('goldens/with_dividers_example/${variant.name}_open_menu.png'),
+        matchesGoldenFile('goldens/with_separators_example/${variant.name}_open_menu.png'),
       );
     },
   );